+++ CHANGELOG.md
@@ -3,14 +3,35 @@
 ## Unreleased
 
 ### API-Changes
-- jsonrpc: add `mailingListAddress` property to `FullChat` #3607
 
 ### Changes
+
+### Fixes
+- share stock string translations across accounts created by the same account manager #3640
+- suppress welcome device messages after account import #3642
+
+## 1.96.0
+
+### Changes
+- jsonrpc js client:
+  - Change package name from `deltachat-jsonrpc-client` to `@deltachat/jsonrpc-client`
+  - remove relative file dependency to it from `deltachat-node` (because it did not work anyway and broke the nix build of desktop)
+  - ci: add github ci action to upload it to our download server automaticaly on realease
+
+## 1.95.0
+
+### API-Changes
+- jsonrpc: add `mailingListAddress` property to `FullChat` #3607
+- jsonrpc: add `MessageNotificationInfo` & `messageGetNotificationInfo()` #3614
+- jsonrpc: add `chat_get_neighboring_media` function #3610
+
+### Changes
+- added `dclogin:` scheme to allow configuration from a qr code
+  (data inside qrcode, contrary to `dcaccount:` which points to an API to create an account) #3541
 - truncate incoming messages by lines instead of just length #3480
 - emit separate `DC_EVENT_MSGS_CHANGED` for each expired message,
   and `DC_EVENT_WEBXDC_INSTANCE_DELETED` when a message contains a webxdc #3605
-
-### Fixes
+- enable `bcc_self` by default #3612
 
 
 ## 1.94.0
@@ -24,7 +45,7 @@
   and `dc_event_emitter_unref()` should be used instead of
   `dc_accounts_event_emitter_unref`.
 - add `dc_contact_was_seen_recently()` #3560
-- Fix get_connectivity_html and get_encrinfo futures not being Send. See rust-lang/rust#101650 for more information
+- Fix `get_connectivity_html` and `get_encrinfo` futures not being Send. See rust-lang/rust#101650 for more information
 - jsonrpc: add functions: #3586, #3587, #3590
   - `deleteChat()`
   - `getChatEncryptionInfo()`